Make the degrees accepting double value instead of integer

+++ lib/src/models/wind/degrees.dart
@@ -1,12 +1,12 @@
 import 'dart:math';
 
-extension type Degrees(int value) {
+extension type Degrees(double value) {
   bool get isValid => value >= 0 && value <= 360;
   Degrees get remainder => Degrees(value % 360);
   double toRadians() => 2 * (value / 360) * pi;
 
   static Degrees fromJson(String json) {
-    return Degrees(int.parse(json));
+    return Degrees(double.parse(json));
   }
 
   String toJson() => staticToJson(this);
